refactor: useTokenData hook
